backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2025.2.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
loguru==0.7.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.1.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pycparser==2.22
pyDatalog==0.17.4
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
supervision==0.25.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.82
ultralytics-thop==2.0.14
urllib3==2.2.2
virtualenv==20.25.0
